Move messaging code out of `@/common`
