+++ src/messaging/protocol.ts
@@ -17,6 +17,13 @@
 
 import { ActionType, Message, SerializedError, Meta } from "@/core";
 import { serializeError } from "serialize-error";
+import { createSendScriptMessage } from "./chrome";
+import {
+  DETECT_FRAMEWORK_VERSIONS,
+  FrameworkMeta,
+  READ_WINDOW,
+  SEARCH_WINDOW,
+} from "./constants";
 
 // eslint-disable-next-line @typescript-eslint/ban-types -- Line can be dropped once we migrate to `webext-messenger`
 export type SerializableResponse = boolean | string | number | object | void;
@@ -64,3 +71,19 @@ export function isRemoteProcedureCallRequest(
   // eslint-disable-next-line @typescript-eslint/no-explicit-any -- This is a type guard function and it uses ?.
   return typeof (message as any)?.type === "string";
 }
+
+type ReadSpec = <T extends Record<string, string>>(arg: {
+  pathSpec: T;
+  waitMillis?: number;
+}) => Promise<Record<keyof T, unknown>>;
+
+export const withReadWindow = createSendScriptMessage(
+  READ_WINDOW
+) as unknown as ReadSpec;
+
+export const withSearchWindow =
+  createSendScriptMessage<{ results: unknown[] }>(SEARCH_WINDOW);
+
+export const withDetectFrameworkVersions = createSendScriptMessage<
+  FrameworkMeta[]
+>(DETECT_FRAMEWORK_VERSIONS);
